refactor: 사이드바 라벨 관련 컴포넌트 리팩토링 (#80)
